Reformatting final tutorial
Rearranged sections to make it more connected - each section broken down by the aspects of the Data Lifecycle. Added highlighted sections for areas we need to address. Included broader discussion on the topic at the top to motivate our research and included Culotta's comments on Milestone 1 so we'd remember to address them.